test(backend): extract shared PDF fixture support
